refactor: extract processor and runner
