+++ 5.38.0/index.ts
@@ -1,6 +1,6 @@
-import { createMorphProject, prettierFormat, yarnInstall } from "../utils";
+import { prettierFormat, runProcessors, yarnInstall } from "../utils";
 import { setupFiles } from "./setupFiles";
-import { Context, IFiles, IProcessor } from "../types";
+import { Context, IProcessor } from "../types";
 import { updateApiSecurityPlugins } from "./updateApiSecurityPlugins";
 import { backupFormLayoutsFolder } from "./backupFormLayoutsFolder";
 import { updateDefaultFormLayout } from "./updateDefaultFormLayout";
@@ -23,27 +23,4 @@ module.exports = async (context: Context) => {
 
     // Install dependencies.
     await yarnInstall();
-};
-
-// Run processors in sequence.
-const runProcessors = async (files: IFiles, processors: IProcessor[], context: Context) => {
-    for (let i = 0; i < processors.length; i++) {
-        const processor = processors[i];
-
-        const rawFiles = files.paths();
-        const project = createMorphProject(rawFiles);
-
-        const result = await processor({
-            context,
-            project,
-            files
-        });
-
-        if (result && result?.skipped !== true) {
-            console.log();
-        }
-
-        // Save file changes.
-        await project.save();
-    }
 };
