+++ utils/processors/createProcessor.ts
@@ -1,0 +1,5 @@
+import {IProcessor} from "../../types";
+
+export const createProcessor = <T = Record<string, any>>(processor: IProcessor<T>) => {
+    return processor;
+};
